[UPD] Moved all of Redux JS to folder "store"
