+++ app/store/configureStore.prod.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { createHashHistory } from 'history';
 import { routerMiddleware } from 'connected-react-router';
-import createRootReducer from '../reducers';
+import createRootReducer from './reducers';
 
 const history = createHashHistory();
 const rootReducer = createRootReducer(history);
