+++ app/store/reducers/app.js
@@ -5,18 +5,18 @@ import { TOGGLE_WORK_ENV, TOGGLE_SETTINGS_WIN } from '../actions/app';
 // Initial state of app
 const initialState = {
   workDevDisabled: true,
-  settingsWinVisible: false
+  settingsWinVisible: false,
 };
 
 export default function appMain(state = initialState, action) {
   switch (action.type) {
     case TOGGLE_WORK_ENV:
       return Object.assign({}, state, {
-        workDevDisabled: action.state
+        workDevDisabled: action.state,
       });
     case TOGGLE_SETTINGS_WIN:
       return Object.assign({}, state, {
-        settingsWinVisible: action.state
+        settingsWinVisible: action.state,
       });
     default:
       return state;
